Modification création Producteur + fixtures + base twig

+++ src/Controller/ProducteurController.php
@@ -2,15 +2,17 @@
 
 namespace App\Controller;
 
-use App\Entity\Categorie;
 use App\Entity\Producteur;
+use App\Entity\User;
 use App\Form\ProducteurFormType;
+use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 
 class ProducteurController extends AbstractController
 {
@@ -32,25 +34,51 @@ class ProducteurController extends AbstractController
      * @Route("/producteur/add", name="creation_producteur")
      * @param Request $request
      * @param EntityManagerInterface $em
+     * @param UserPasswordEncoderInterface $passwordEncoder
      * @return RedirectResponse|Response
      */
-    public function add(Request $request, EntityManagerInterface $em)
+    public function add(Request $request, EntityManagerInterface $em, UserPasswordEncoderInterface $passwordEncoder)
     {
-        $producteur = new Producteur();
-        $form = $this->createForm(ProducteurFormType::class, $producteur);
-        $form->handleRequest($request);
+        if ($this->isGranted('ROLE_ADMIN')) {
 
-        if ($form->isSubmitted() && $form->isValid()) {
+            $producteur = new Producteur();
+            $form = $this->createForm(ProducteurFormType::class, $producteur);
+            $form->handleRequest($request);
 
-            $em->persist($producteur);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
 
-            return $this->redirectToRoute("producteur");
+                $user = new User();
+                $user->setNom($form->get("nom")->getData());
+                $user->setPrenom($form->get("prenom")->getData());
+                $user->setAdresse($form->get("adresse")->getData());
+                $user->setCodePostal($form->get("codePostal")->getData());
+                $user->setRoles(['ROLE_PRODUCTEUR']);
+                $user->setDateInscription(new DateTime('now'));
+                $user->setEmail($form->get("email")->getData());
+                $user->setVille($form->get("ville")->getData());
+                $user->setTelephone($form->get("telephone")->getData());
+                $user->setValide(1);
+                $user->setPassword(
+                    $passwordEncoder->encodePassword(
+                        $user,
+                        $form->get('plainPassword')->getData()
+                    )
+                );
+                $em->persist($user);
+                $em->persist($producteur);
+                $em->flush();
+
+                return $this->redirectToRoute("producteur");
+            }
+
+            return $this->render('producteur/creationProducteur.html.twig', [
+                    'producteurForm' => $form->createView()]
+            );
+
+        } else {
+            $this->addFlash("error", "Droit d'administrateur requis !");
+            return $this->redirectToRoute("home");
         }
-
-        return $this->render('producteur/creationProducteur.html.twig', [
-                'producteurForm' => $form->createView()]
-        );
     }
 
     /**
@@ -77,23 +105,28 @@ class ProducteurController extends AbstractController
      */
     public function edit($id, Request $request, EntityManagerInterface $em)
     {
-        $producteurRepo = $this->getDoctrine()->getRepository(Producteur::class);
-        $producteur = $producteurRepo->find($id);
-        $form = $this->createForm(ProducteurFormType::class, $producteur);
-        $form->handleRequest($request);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $producteurRepo = $this->getDoctrine()->getRepository(Producteur::class);
+            $producteur = $producteurRepo->find($id);
+            $form = $this->createForm(ProducteurFormType::class, $producteur);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+            if ($form->isSubmitted() && $form->isValid()) {
 
-            $em->persist($producteur);
-            $em->flush();
+                $em->persist($producteur);
+                $em->flush();
 
-            return $this->redirectToRoute("producteur");
+                return $this->redirectToRoute("producteur");
+            }
+
+            return $this->render('producteur/edit.html.twig', [
+                    'producteurForm' => $form->createView(),
+                    'id' => $id]
+            );
+        } else {
+            $this->addFlash("error", "Droit d'administrateur requis !");
+            return $this->redirectToRoute("home");
         }
-
-        return $this->render('producteur/edit.html.twig', [
-                'producteurForm' => $form->createView(),
-                'id' => $id]
-        );
     }
 
     /**
@@ -104,13 +137,18 @@ class ProducteurController extends AbstractController
      */
     public function delete($id, EntityManagerInterface $em)
     {
-        $producteurRepo = $this->getDoctrine()->getRepository(Producteur::class);
-        $producteur = $producteurRepo->find($id);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $producteurRepo = $this->getDoctrine()->getRepository(Producteur::class);
+            $producteur = $producteurRepo->find($id);
 
-        if ($producteur != null) {
-            $em->remove($producteur);
-            $em->flush();
+            if ($producteur != null) {
+                $em->remove($producteur);
+                $em->flush();
+            }
+            return $this->redirectToRoute("producteur");
+        } else {
+            $this->addFlash("error", "Droit d'administrateur requis !");
+            return $this->redirectToRoute("home");
         }
-        return $this->redirectToRoute("producteur");
     }
 }
